List: consolidate the indent and outdent helpers into a single utils file
Fold the utils folder back into one utils.js, and drop the now redundant
first-item guard and client ID from useSpace: indentListItems already
returns false when there is no sibling to nest under and falls back to
the selection, so the space and tab handler matches the multi-selection one.
