--- packages/block-library/src/list-item/utils/index.js
@@ -1,4 +1,0 @@
-export { getIndentTarget, getOutdentTarget } from './indent-outdent-targets';
-export { moveBlocksToNestedList } from './move-blocks-to-nested-list';
-export { indentListItems } from './indent-list-items';
-export { outdentListItems } from './outdent-list-items';
